fix: properly record the end of the measure and use it
The sampler now records the last update time and provides the start time
for the new update to the sensor. For RAPL, this makes things a little
precise instead of using the sampling period. For powermetrics, we can
extract the measure time from the output, also making things more
precise.